requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
aiosqlite==0.20.0
pyjwt==2.9.0
cryptography==43.0.1